refactor reset methods into owned file
